+++ src/main.rs
@@ -3,6 +3,7 @@
 use std::sync::Arc;
 pub mod commands;
 mod egui_impl;
+pub mod renderer;
 pub mod repositories;
 pub mod vulkano_prelude;
 pub mod window;
@@ -124,262 +125,6 @@ impl GlobalHotkeys {
     }
 }
 
-mod stroke_renderer {
-    /// The data managed by the renderer.
-    /// For now, in persuit of actually getting a working product one day,
-    /// this is a very coarse caching sceme. In the future, perhaps a bit more granular
-    /// control can occur, should performance become an issue:
-    ///  * Caching images of incrementally older states, reducing work to get to any given state (performant undo)
-    ///  * Caching tesselation output
-    pub struct RenderData {
-        image: Arc<vk::StorageImage>,
-        pub view: Arc<vk::ImageView<vk::StorageImage>>,
-    }
-
-    use crate::vk;
-    use anyhow::Result as AnyResult;
-    use std::sync::Arc;
-    use vulkano::{pipeline::graphics::vertex_input::Vertex, pipeline::Pipeline, sync::GpuFuture};
-    mod vert {
-        vulkano_shaders::shader! {
-            ty: "vertex",
-            path: "src/shaders/stamp.vert",
-        }
-    }
-    mod frag {
-        vulkano_shaders::shader! {
-            ty: "fragment",
-            path: "src/shaders/stamp.frag",
-        }
-    }
-
-    pub struct StrokeLayerRenderer {
-        context: Arc<crate::render_device::RenderContext>,
-        texture_descriptor: Arc<vk::PersistentDescriptorSet>,
-        gpu_tess: super::gpu_tess::GpuStampTess,
-        pipeline: Arc<vk::GraphicsPipeline>,
-    }
-    impl StrokeLayerRenderer {
-        pub fn new(context: Arc<crate::render_device::RenderContext>) -> AnyResult<Self> {
-            let image = image::open("brushes/splotch.png")
-                .unwrap()
-                .into_luma_alpha8();
-
-            //Iter over transparencies.
-            let image_grey = image.iter().skip(1).step_by(2).cloned();
-
-            let mut cb = vk::AutoCommandBufferBuilder::primary(
-                context.allocators().command_buffer(),
-                context.queues().transfer().idx(),
-                vulkano::command_buffer::CommandBufferUsage::OneTimeSubmit,
-            )?;
-            let (image, sampler) = {
-                let image = vk::ImmutableImage::from_iter(
-                    context.allocators().memory(),
-                    image_grey,
-                    vk::ImageDimensions::Dim2d {
-                        width: image.width(),
-                        height: image.height(),
-                        array_layers: 1,
-                    },
-                    vulkano::image::MipmapsCount::One,
-                    vk::Format::R8_UNORM,
-                    &mut cb,
-                )?;
-                context
-                    .now()
-                    .then_execute(context.queues().transfer().queue().clone(), cb.build()?)?
-                    .then_signal_fence_and_flush()?
-                    .wait(None)?;
-
-                let view = vk::ImageView::new(
-                    image.clone(),
-                    vk::ImageViewCreateInfo {
-                        component_mapping: vk::ComponentMapping {
-                            //Red is coverage of white, with premul.
-                            a: vk::ComponentSwizzle::Red,
-                            r: vk::ComponentSwizzle::Red,
-                            b: vk::ComponentSwizzle::Red,
-                            g: vk::ComponentSwizzle::Red,
-                        },
-                        ..vk::ImageViewCreateInfo::from_image(&image)
-                    },
-                )?;
-
-                let sampler = vk::Sampler::new(
-                    context.device().clone(),
-                    vk::SamplerCreateInfo {
-                        min_filter: vk::Filter::Linear,
-                        mag_filter: vk::Filter::Linear,
-                        ..Default::default()
-                    },
-                )?;
-
-                (view, sampler)
-            };
-
-            let frag = frag::load(context.device().clone())?;
-            let vert = vert::load(context.device().clone())?;
-            // Unwraps ok here, using GLSL where "main" is the only allowed entry point.
-            let frag = frag.entry_point("main").unwrap();
-            let vert = vert.entry_point("main").unwrap();
-
-            // DualSrcBlend (~75% coverage) is used to control whether to erase or draw on a per-fragment basis
-            // [1.0; 4] = draw, [0.0; 4] = erase.
-            let mut premul_dyn_constants = vk::ColorBlendState::new(1);
-            premul_dyn_constants.blend_constants = vk::StateMode::Fixed([1.0; 4]);
-            premul_dyn_constants.attachments[0].blend = Some(vk::AttachmentBlend {
-                alpha_source: vulkano::pipeline::graphics::color_blend::BlendFactor::Src1Alpha,
-                color_source: vulkano::pipeline::graphics::color_blend::BlendFactor::Src1Color,
-                alpha_destination:
-                    vulkano::pipeline::graphics::color_blend::BlendFactor::OneMinusSrcAlpha,
-                color_destination:
-                    vulkano::pipeline::graphics::color_blend::BlendFactor::OneMinusSrcAlpha,
-                alpha_op: vulkano::pipeline::graphics::color_blend::BlendOp::Add,
-                color_op: vulkano::pipeline::graphics::color_blend::BlendOp::Add,
-            });
-
-            let pipeline = vk::GraphicsPipeline::start()
-                .fragment_shader(frag, ())
-                .vertex_shader(vert, ())
-                .vertex_input_state(super::gpu_tess::interface::OutputStrokeVertex::per_vertex())
-                .input_assembly_state(vk::InputAssemblyState::new()) //Triangle list, no prim restart
-                .color_blend_state(premul_dyn_constants)
-                .rasterization_state(vk::RasterizationState::new()) // No cull
-                .viewport_state(vk::ViewportState::viewport_fixed_scissor_irrelevant([
-                    vk::Viewport {
-                        depth_range: 0.0..1.0,
-                        dimensions: [super::DOCUMENT_DIMENSION as f32; 2],
-                        origin: [0.0; 2],
-                    },
-                ]))
-                .render_pass(
-                    vulkano::pipeline::graphics::render_pass::PipelineRenderPassType::BeginRendering(
-                        vulkano::pipeline::graphics::render_pass::PipelineRenderingCreateInfo {
-                            view_mask: 0,
-                            color_attachment_formats: vec![Some(super::DOCUMENT_FORMAT)],
-                            depth_attachment_format: None,
-                            stencil_attachment_format: None,
-                            ..Default::default()
-                        }
-                    )
-                )
-                .build(context.device().clone())?;
-
-            let descriptor_set = vk::PersistentDescriptorSet::new(
-                context.allocators().descriptor_set(),
-                pipeline.layout().set_layouts()[0].clone(),
-                [vk::WriteDescriptorSet::image_view_sampler(
-                    0, image, sampler,
-                )],
-            )?;
-
-            let tess = super::gpu_tess::GpuStampTess::new(context.clone())?;
-
-            Ok(Self {
-                context,
-                pipeline,
-                gpu_tess: tess,
-                texture_descriptor: descriptor_set,
-            })
-        }
-        /// Allocate a new RenderData object. Initial contents are undefined!
-        pub fn uninit_render_data(&self) -> anyhow::Result<RenderData> {
-            let image = vk::StorageImage::with_usage(
-                self.context.allocators().memory(),
-                vulkano::image::ImageDimensions::Dim2d {
-                    width: super::DOCUMENT_DIMENSION,
-                    height: super::DOCUMENT_DIMENSION,
-                    array_layers: 1,
-                },
-                super::DOCUMENT_FORMAT,
-                vk::ImageUsage::COLOR_ATTACHMENT | vk::ImageUsage::STORAGE,
-                vk::ImageCreateFlags::empty(),
-                [
-                    // Todo: if these are the same queue, what happen?
-                    self.context.queues().graphics().idx(),
-                    self.context.queues().compute().idx(),
-                ]
-                .into_iter(),
-            )?;
-            let view = vk::ImageView::new_default(image.clone())?;
-
-            use vulkano::VulkanObject;
-            log::info!("Made render data at id{:?}", view.handle());
-
-            Ok(RenderData { image, view })
-        }
-        pub fn draw(
-            &self,
-            strokes: &[crate::state::stroke_collection::ImmutableStroke],
-            renderbuf: &RenderData,
-            clear: bool,
-        ) -> AnyResult<vk::sync::future::SemaphoreSignalFuture<impl vk::sync::GpuFuture>> {
-            let (future, vertices, indirects) = self.gpu_tess.tess(strokes)?;
-            let mut command_buffer = vk::AutoCommandBufferBuilder::primary(
-                self.context.allocators().command_buffer(),
-                self.context.queues().graphics().idx(),
-                vulkano::command_buffer::CommandBufferUsage::OneTimeSubmit,
-            )?;
-
-            let mut matrix = cgmath::Matrix4::from_scale(2.0 / super::DOCUMENT_DIMENSION as f32);
-            matrix.y *= -1.0;
-            matrix.w.x -= 1.0;
-            matrix.w.y += 1.0;
-
-            command_buffer
-                .begin_rendering(vulkano::command_buffer::RenderingInfo {
-                    color_attachments: vec![Some(
-                        vulkano::command_buffer::RenderingAttachmentInfo {
-                            clear_value: if clear {
-                                Some([0.0, 0.0, 0.0, 0.0].into())
-                            } else {
-                                None
-                            },
-                            load_op: if clear {
-                                vulkano::render_pass::LoadOp::Clear
-                            } else {
-                                vulkano::render_pass::LoadOp::Load
-                            },
-                            store_op: vulkano::render_pass::StoreOp::Store,
-                            ..vulkano::command_buffer::RenderingAttachmentInfo::image_view(
-                                renderbuf.view.clone(),
-                            )
-                        },
-                    )],
-                    contents: vulkano::command_buffer::SubpassContents::Inline,
-                    depth_attachment: None,
-                    ..Default::default()
-                })?
-                .bind_pipeline_graphics(self.pipeline.clone())
-                .push_constants(
-                    self.pipeline.layout().clone(),
-                    0,
-                    Into::<[[f32; 4]; 4]>::into(matrix),
-                )
-                .bind_descriptor_sets(
-                    vulkano::pipeline::PipelineBindPoint::Graphics,
-                    self.pipeline.layout().clone(),
-                    0,
-                    self.texture_descriptor.clone(),
-                )
-                .bind_vertex_buffers(0, vertices)
-                .draw_indirect(indirects)?
-                .end_rendering()?;
-
-            let command_buffer = command_buffer.build()?;
-
-            // After tessellation finishes, render.
-            Ok(future
-                .then_execute(
-                    self.context.queues().graphics().queue().clone(),
-                    command_buffer,
-                )?
-                .then_signal_semaphore_and_flush()?)
-        }
-    }
-}
-
 #[derive(bytemuck::Pod, bytemuck::Zeroable, Clone, Copy)]
 #[repr(C)]
 pub struct StrokePoint {
@@ -409,280 +154,6 @@ pub struct Stroke {
     brush: state::StrokeBrushSettings,
     points: Vec<StrokePoint>,
 }
-async fn render_worker(
-    renderer: Arc<render_device::RenderContext>,
-    document_preview: Arc<document_viewport_proxy::DocumentViewportPreviewProxy>,
-    mut render_recv: tokio::sync::mpsc::UnboundedReceiver<()>,
-) -> AnyResult<()> {
-    /*
-    let layer_render = stroke_renderer::StrokeLayerRenderer::new(renderer.clone())?;
-    let blend_engine = blend::BlendEngine::new(renderer.device().clone())?;
-
-    let globals = GLOBALS.get_or_init(Globals::new);
-
-    let mut cached_renders =
-        std::collections::HashMap::<FuzzID<StrokeLayer>, stroke_renderer::RenderData>::new();
-    // Keep track of layer states locally, as globals can be mutated out-of-step with render commands
-    // We expect eventual consistency though!
-    let mut weak_layer_strokes =
-        std::collections::HashMap::<FuzzID<StrokeLayer>, Vec<ImmutableStroke>>::new();
-
-    // An event that was peeked and rejected during aggregation.
-    // Take it the next time around.
-    let mut peeked = None;
-
-    let mut document_to_draw = None;
-
-    // Take the peeked value, or try to receive new value.
-    while let Some(message) = match peeked.take() {
-        Some(v) => {
-            // We skip awaiting the reciever, and have more work immediately.
-            // yield so that we don't starve the runtime.
-            tokio::task::yield_now().await;
-            Some(v)
-        }
-        None => render_recv.recv().await,
-    } {
-        // Try to aggregate more events, to batch work effectively.
-        let mut events = vec![message];
-        let mut skip_blend = false;
-        loop {
-            match render_recv.try_recv() {
-                Ok(v) => {
-                    // Accept value, or put into `peeked` and break if it
-                    // is incompatible with aggregated work
-
-                    // unwrap ok - vec always has at least one element.
-                    let can_aggregate = match (events.last().unwrap(), &v) {
-                        // Multiple edits to the same layer can be aggregated
-                        (
-                            RenderMessage::StrokeLayer { layer: target, .. },
-                            RenderMessage::StrokeLayer { layer: new, .. },
-                        ) if new == target => true,
-                        // Multiple document switches can be aggregated
-                        // (all are ignored but the last one)
-                        (RenderMessage::SwitchDocument(..), RenderMessage::SwitchDocument(..)) => {
-                            true
-                        }
-                        // Defer switch to next pass, but the switching document means we can skip
-                        // re-rendering the document after the stroke layer commands.
-                        (RenderMessage::StrokeLayer { .. }, RenderMessage::SwitchDocument(..)) => {
-                            skip_blend = true;
-                            false
-                        }
-                        _ => {
-                            // Incompatible events. Defer event to next pass, and break
-                            false
-                        }
-                    };
-
-                    if can_aggregate {
-                        events.push(v);
-                    } else {
-                        // peeked is always None here, as we took it above. no events lost :3
-                        peeked = Some(v);
-                        break;
-                    }
-                }
-                Err(_) => {
-                    // Empty or closed, we don't care - break and work
-                    // on what events we managed to collect (always at least one)
-                    break;
-                }
-            }
-        }
-        match events.first().unwrap() {
-            // All our commands are layer commands relating to this target:
-            RenderMessage::StrokeLayer { layer: target, .. } => {
-                // clear the layer image before rendering? true if undone past the cached version.
-                let mut clear = false;
-                // keep track of rendering work to do. May include old strokes too,
-                // if the truncate commands put us back in time before the cached image.
-                let mut strokes_to_render = Vec::<ImmutableStroke>::new();
-                let layer_strokes = weak_layer_strokes.entry(target.clone()).or_default();
-
-                // Keep track of the blend changes. Only the last will apply.
-                // None at the end of iteration means it hasn't changed - read from
-                // document data directly.
-                let mut new_blend = None;
-                for event in events.iter() {
-                    match event {
-                        RenderMessage::StrokeLayer { layer, kind } => {
-                            assert!(layer == target);
-
-                            match kind {
-                                StrokeLayerRenderMessageKind::Append(s) => {
-                                    strokes_to_render.push(s.clone());
-                                    layer_strokes.push(s.clone());
-                                }
-                                StrokeLayerRenderMessageKind::Truncate(num) => {
-                                    let num_to_render = strokes_to_render.len();
-
-                                    // Remove strokes from render queue
-                                    strokes_to_render.drain(num_to_render.saturating_sub(*num)..);
-
-                                    let global_strokes_truncate = layer_strokes
-                                        .len()
-                                        .checked_sub(*num)
-                                        .expect("Cant truncate past empty");
-
-                                    layer_strokes.drain(global_strokes_truncate..);
-
-                                    // we took as many as possible from new commands - how many left over to remove?
-                                    // if more than zero, we have to fetch strokes from globals.
-                                    let num = num.saturating_sub(num_to_render);
-                                    if num > 0 {
-                                        // We're no longer strictly adding new content, we're removing what's
-                                        // already been drawn. Clear the image.
-                                        clear = true;
-                                        strokes_to_render.extend(layer_strokes.iter().cloned());
-                                    }
-                                }
-                                StrokeLayerRenderMessageKind::BlendChanged(blend) => {
-                                    new_blend = Some(blend);
-                                }
-                            }
-                        }
-                        _ => panic!("Incorrect render command aggregation!"),
-                    }
-                }
-
-                if !clear && strokes_to_render.is_empty() && new_blend.is_none() {
-                    // No work to do. Go listen for new events.
-                    continue;
-                }
-
-                let mut draw_semaphore_future = None;
-
-                if clear && strokes_to_render.len() == 0 {
-                    // Clear without remaking - just delete the data.
-                    cached_renders.remove(&target);
-                } else {
-                    // get or try insert with - create buffer if it doesn't exist, and get a ref to it.
-                    let cache_entry = match cached_renders.entry(target.clone()) {
-                        std::collections::hash_map::Entry::Occupied(occupied) => {
-                            &*occupied.into_mut()
-                        }
-                        std::collections::hash_map::Entry::Vacant(vacant) => {
-                            &*vacant.insert(layer_render.uninit_render_data()?)
-                        }
-                    };
-                    match layer_render.draw(&strokes_to_render, cache_entry, clear) {
-                        Ok(semaphore) => draw_semaphore_future = Some(semaphore),
-                        Err(e) => {
-                            log::warn!("{e:?}");
-                        }
-                    };
-                }
-
-                // implicit sync: if skip_blend or document_to_draw.is_none the draw_semaphore_future will not be awaited.
-                // next draw to this layer may need to be synchronized. Vulkano handles this in the background, but
-                // it's an important subtlety so... note for future self!
-
-                if !skip_blend {
-                    if let Some(document_to_draw) = document_to_draw {
-                        // wawawawawaaaw
-                        // let blend = new_blend.unwrap();
-                        // collect blend info from realtime document.
-                        // need to track this locally, this info could be from wayyy in the future.
-                        let blend_info: Vec<_> = {
-                            let read = globals.documents.read().await;
-                            let Some(doc) = read.iter().find(|doc| doc.id == document_to_draw)
-                            else {
-                                // Document not found, nothin to draw.
-                                continue;
-                            };
-
-                            doc.layer_top_level
-                                .iter()
-                                .filter_map(|layer| {
-                                    Some((
-                                        layer.blend.clone(),
-                                        cached_renders.get(&layer.id)?.view.clone(),
-                                    ))
-                                })
-                                .collect()
-                        };
-
-                        let proxy = document_preview.write().await;
-                        let commands = blend_engine.blend(
-                            &renderer,
-                            proxy.clone(),
-                            true,
-                            &blend_info,
-                            [0; 2],
-                            [0; 2],
-                        )?;
-
-                        let fence = match draw_semaphore_future {
-                            Some(semaphore) => semaphore
-                                .then_execute(
-                                    renderer.queues().compute().queue().clone(),
-                                    commands,
-                                )?
-                                .boxed_send(),
-                            None => renderer
-                                .now()
-                                .then_execute(
-                                    renderer.queues().compute().queue().clone(),
-                                    commands,
-                                )?
-                                .boxed_send(),
-                        }
-                        .then_signal_fence_and_flush()?;
-                        proxy.submit_with_fence(fence);
-                    }
-                }
-            }
-            // All our commands are document switch commands:
-            // (we only care about the final one)
-            RenderMessage::SwitchDocument(..) => {
-                let RenderMessage::SwitchDocument(new_document) = events.last().unwrap() else {
-                    panic!("Incorrect render command aggregation!")
-                };
-                document_to_draw = Some(new_document.clone());
-
-                // <rerender viewport>
-                let blend_info: Vec<_> = {
-                    let read = globals.documents.read().await;
-                    let Some(doc) = read.iter().find(|doc| doc.id == *new_document) else {
-                        // Document not found, nothin to draw.
-                        continue;
-                    };
-
-                    doc.layer_top_level
-                        .iter()
-                        .filter_map(|layer| {
-                            Some((
-                                layer.blend.clone(),
-                                cached_renders.get(&layer.id)?.view.clone(),
-                            ))
-                        })
-                        .collect()
-                };
-
-                let proxy = document_preview.write().await;
-                let commands = blend_engine.blend(
-                    &renderer,
-                    proxy.clone(),
-                    true,
-                    &blend_info,
-                    [0; 2],
-                    [0; 2],
-                )?;
-
-                let fence = renderer
-                    .now()
-                    .then_execute(renderer.queues().compute().queue().clone(), commands)?
-                    .boxed_send()
-                    .then_signal_fence_and_flush()?;
-                proxy.submit_with_fence(fence);
-            }
-        }
-    }*/
-    while render_recv.recv().await.is_some() {}
-    Ok(())
-}
 async fn stylus_event_collector(
     mut event_stream: tokio::sync::broadcast::Receiver<stylus_events::StylusEventFrame>,
     mut action_listener: actions::ActionListener,
@@ -690,31 +161,6 @@ async fn stylus_event_collector(
     document_preview: Arc<document_viewport_proxy::DocumentViewportPreviewProxy>,
     render_send: tokio::sync::mpsc::UnboundedSender<()>,
 ) -> AnyResult<()> {
-    // Create a document and a few layers and select them, to speed up testing iterations :P
-    /*
-    let globals = GLOBALS.get_or_init(Globals::new);
-    {
-        let (default_document, default_layer) = {
-            let mut document = Document::default();
-            let document_id = document.id.weak();
-
-            let layer = StrokeLayer::default();
-            let layer_id = layer.id.weak();
-            document.layer_top_level.push(layer);
-
-            globals.documents.write().await.push(document);
-            (document_id, layer_id)
-        };
-
-        let mut selections = globals.selections().write().await;
-        selections.cur_document = Some(default_document);
-        let document_selections = selections
-            .document_selections
-            .entry(default_document)
-            .or_default();
-        document_selections.cur_layer = Some(default_layer);
-    }*/
-
     loop {
         match event_stream.recv().await {
             Ok(stylus_frame) => {
@@ -822,7 +268,7 @@ fn main() -> AnyResult<std::convert::Infallible> {
             // for now, just a note for future self UwU
             runtime.block_on(async {
                 tokio::try_join!(
-                    render_worker(render_context, document_view.clone(), render_reciever,),
+                    renderer::render_worker(render_context, document_view.clone(), render_reciever,),
                     stylus_event_collector(
                         event_stream,
                         action_listener,
